Fixed Measure Notes Security finding for Deleting Measure Notes.
git-svn-id: svn+ssh://subversion.telligen.com/var/www/html/repos/cmat/trunk/Development/mat@11869 de2f59c9-771b-4d1d-97da-b950f46c8601

Former-commit-id: ccd7e4e6ab5033ddfa5988bf67904735a0fe1800

+++ src/mat/server/MeasureLibraryServiceImpl.java
@@ -980,7 +980,10 @@ public class MeasureLibraryServiceImpl implements MeasureLibraryService {
 		MeasureNotesDAO measureNotesDAO = getMeasureNotesDAO();
 		MeasureNotes measureNotes = measureNotesDAO.find(measureNoteDTO.getId());
 		try {
-			getMeasureNotesService().deleteMeasureNote(measureNotes);
+			if(MatContextServiceUtil.get().isCurrentMeasureEditable(getMeasureDAO(),measureNoteDTO.getMeasureId())){
+				getMeasureNotesService().deleteMeasureNote(measureNotes);
+			}
+			
 			logger.info("MeasureNotes Deleted Successfully :: " + measureNotes.getId());
 		} catch (Exception e) {
 			logger.info("MeasureNotes not deleted. Exception occured. Measure notes Id :: " + measureNotes.getId());
